Remove all Samsung's sdcardfs for a reset
We're still going to use the upstream version of sdcardfs for
some obvious reasons, like maintainability.

The upstream sdcardfs has diverged with this version, making it hard to
import changes from the upstream.

So yeah, to use this version again, you have to revert the upstream and
revert this commit.