Re-include Samsung's sdcardfs under a differennt config
There is an incompatability between the newer version of sdcardfs
and the impementation in the official kitkat rom. For example, the mount
options used by the rom is not present anymore and the default action of
the newer driver is to fail. Even if it did mount
(i.e by modifying the code to allow unsupported mount options or redirecting
the options to their equivalents, if there are any), there would still be a lot of
permission problems.

But for the both sdcardfs versions to exist, there's a bit of a renaming
to the the configs. Of course, it's not possible to include the both versions
in a build.

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>