input: ist3038: Touch Booster: Release limits when resetting
When Touch Booster, a.ka Touch Boost, is compiled in, the cpu clocks up
when a touch event occurred then clocks down to normal on release.
This is done by imposing a limit using the api fromcpufreq_limit
to raise the min frequency to the maximum fequency.

As defined in ist30xx.c, the number of fingers pressed down is counted in a
separate static variable which may not be the same as the hardware counters.

Whn the touchscreen resets, the counters are reset to 0 but not that separate variable
which may lead to problems that result to the cpu clocking up to the maximum frequency.

To reproduce:
1. compile a kernel with TOUCH_BOOSTER macro defined to 1 (look in ist30xx.c)
2. Boot as usual.
3. Touch or swipe the phone's screen while it is still on
4. Whitout lifting your finger, connect then disconnect the phone to your computer (doesn't seem to happen in ac adaptor)
5. Then without turning off the screen or the phone, check the fequency of the cpus' clocks

There is another problem, even though the limit is lifted, there should be a case
where the minimum cpufreq frequency will not go back to the value before the limit
is imposed. This patch WILL NOT SOLVE THAT.

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>