cpufreq: sprdemand: Abort functions when sprdemand is not the current governor
Do not assume that sprdemand is always used.

There are functions dependent on sprdemand's private data that run even if it doesn't exists.
Notably, the functions for the pm and thermal callbacks.

The problem is that the private data is extracted from the tuners member of struct dbs_data.
The tuners member given it's a void *; it can be anything.
It can be NULL or the sprdemand private data we wanted or even another governor's private data.
That gives us three cases. The functions do check if it's null but that's all.
It does not guarantee that if it's not NULL the data is the sprdemand's data.
They hope that the user always use sprdemand.

Whenever those functions detect that tuners member is not NULL but does not point to the
sprdemand's private data, and tries to access a member. You would probably
get something like this and the kernel will panic.

	Unable to handle kernel paging request at virtual address XXXXXXXX

The fix is a bit naive as of now and that's to check for the governors name.
That's it.