ThunderCharge: add custom_current file back as a backwards compatibility
With ThunderCharge v2, the custom curent file is split to ac and usb
but the problem is that there are some userspace tools , cough Ker.. cough diut...,
that doesn't support this yet.

So I'll be adding this back.

This custom_current file will set both the ac and usb part at once but
when read should output the current current setting for the ac part

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>