mali-400: sc8830: use core33g sc8830 plaform code; USING_PP_CORE is set to 4
Using a newer revision of the kernel driver would break
the userspace driver thus just simply using this driver is with the
stock/virgin userspace driver is a no-no.

As an experiment why not just import the plaform code which should
implement a different style of dfs?

I've noticed in the stock rom that the mali_max_pp_cores_group_1 always
reads 4 but on the recent builds only reads 1. I think it is from the scx15
platform code that only uses 1 regardless of what USING_PP_CORE is defined.
This code reads the MALI_PP_CORE_NUMBER macro, which is defined from
USING_PP_CORE, and hopefully let us use 4 instead of one, plus a frequency
scaling algorithm.