zram: Fix implicit declaration error to zs_get_total_size_bytes
Due to a custom code, meaning not being able to autommatically be reappropiated
upon an api change via a commit, which still uses zs_get_total_size_bytes
the compilation will not proceed due to an implicit declaration of a function
which is not defined anymore.

The error is due to "zsmalloc: change return value unit of zs_get_total_size_bytes"

Luckily, the function in the custom code adds all the bytes then shifts it to a constant which
would convert it to pages. The above change just modifies the
zs_get_total_size_bytes to become zs_get_total_pages. In other terms, the function
is modified so it would output the value in pages instead of bytes, which is what the
function needs.

Renaming the function call then removing the shift should fix this up.