cpufreq: sprdemand: Only check for NULL in g_sd_tuners instead of comparing the name
This is a less expensive approach taking to account that the g_sd_tuners pointer is
always NULL when sprdemand is not initialized as it inherently points(word: aliases) to
dbs_data->tuners(prase: sprdemand's private data). So the tuners member is not null, we
just check if the g_sd_tuners variable is not NULL. If it is then sprdemand has been initiallized.
If it's not then it's not, and we should abort.

On some versions of this code,
g_sd_tuners points to an allocated area which means this test won't be as reliable as for this one.