cpufreq_sprdemand: Do not use higher values for high thresholds
The "jump to the highest frequency when load is higher than up_threshold parameter"
part has been adjusted so that it only does that when all cores are online.
If not then it only "jumps the frequency to policy->max-1". But that's not usually
the case

By using the default "95" for up_threshold may incur a this case.
For example, assuming we're using a unmodified frequency table:

step 1: 1 core online, load = 87
    target freq should be the nearest frequency greater than 0.87* 1200MHz
    which should be 1200MHz since it's the next higher freq than 1044Hz
step 2: 2 cores online, load = 96
    usually it's 1Ghz, so it went down
step 3: 3 cores online, load = 100
    same as step 2, so it's about 1000Hz again
step 4: all 4 cores now online, load 100
    now it could scale up to 12000Mhz

I think this wass made to conserve battery power, but sometimes
doesn't follow because the up_threshold is too high.
It prevents scaling to maximum but you could easily reach it
even when:
    load < up_threshold
If we were to retain that change, the up_threshold
should be less than this expression.
((policy->max-1)*100)/policy->max.