cpufreq: Assume there's only one policy when suspending governors
A commit "cpufreq: suspend governors on system suspend/hibernate"
was backported and relies upon that the policies are accessible
via a list. in Linux 3.10, that's not yet implemented.

Not that it's not easy to implement, it just alters the api which may
break compatibility with other cpufreq governors out there.

So for now, assume that there's only one policy used for all cpus.
Yes, this is not usually the case. Try to examine the device
under LPM(that mode where you charge the device when its powered down
and it shows a little animation of a charging battery) and you should
see that cpu0 has been set to powersave and the other cores to the
default governor. It's in the default initrd and should be fixable.