arm: mach-sc: Increase W modem memory [WIP]
This is based on commit from:
https://github.com/remilia15/android_kernel_samsung_core33g/commit/699735934faf62d996fba08a36e7c561e5f577c

Running Dual sim Dual standby would result to an assertion error after a couple of hours
causing the device to reboot, to trigger a kernel panic,
or to shut down the telephony services and daemons and restart them when possible.

The behaviour depends on which modemd binary you use and the set properties.

Testing for issue remilia15/android_device_samsung_scx30g2-common#1