kanas: add impasta_kanas_cm14_1_defconfig for cm_14.1 kanas builds
The YAC(Yet Another Custom) Kernel is somewhat an unreleased project that is born from
both OM and Sandroid kernels.

I tried working from OM first but it didn't work on stock rom. Then, I reworked it based on
Sandroid to make it work. Finally, I reworked it again but now based on samsung opensource zips.

I plan to make a PR or release it as a new project back in 2017
But since I was afraid of the responsibilities I didn't and gradually forgot about this.

Still, I'm not going to release it. :P

To make maintenance simpler, there is only a yack_testing branch.
I made it so to support stock till android parfait(?). I haven't tested it though.
It just needs different configs for each android versions.

This adds the defconfig file for the cm-14.1 kanas3g builds, if there are any.

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>