BMA254: add Significant Motion Sensor support [WIP]
BOSCH's bma255 sensor seems to have support for SMD like what has been written
on their bma2x2 driver. Now, kanas's bma sensor seems to be based on bma255
if not a real bma255 sensor since Samsung did wrote their own driver.
And yes, bma2x2 driver will not work since it fails to probe the chip.

Because of the presence of the reactive alert flag, it gave me confidence that
the significant motion detection part of the bma2x2 driver can be cross ported to
the Samsung one.

If this experiment succeeds, the software implementation of SMD by polling
the accelerometer 30 times then computing the linear acceleration finally comparing to
an arbitrary hardcoded threshold to determine motion is not anymore needed.

Man that's really long, fairly complicated, and slow.

And, Android Doze can be fully utilized.
Well, that is only if one can write or reverse-engineer
the Android HAL driver. :)

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>