arm: mach-sc: increase cpufreq table capacity
Due to this:
mach-sc: Allow more flexible changing of the frequency and voltage table for the processor

It is that much possible to create a buffer overflow by activating all options.
The effect of the buffer overflow should vary but one thing for sure that it will mess up the
cpu frequency table and would very likely to cause kernel panics due to some illegal access.
From my observations, it's almost certain that the system will panic when scaling down to the
lowest set frequency in the table. That is, when the system goes idle.

Increase it so it won't happen, ever again.

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>