config-14.1: Update config to reflect current state of the kernel
Some new cpu hotplugging drivers are added (alucard, autosmp and intelli_plug) and
I wrote an interface to coordinate these drivers (and it's complete with documentation too)

Thundercharger is also enabled. Power-efficient workqueues are added and enabled.

Frontswap is never used so it was removed.

The timer frequency is back to 100Hz to explore the mysterious queue is filled assertion error in the coprocessor
that happends at least twice every month.

Lastly, we're back to interactive being default cpufreq governor.

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>