BMA254: Use the SMD thresholds from the bma2x2 driver
This should make the SMD more sensistive as before. It seems that it requires
a siginifcant motion to occur within 3 seconds with half the sensitivity as before.

There's no reason why to write this patch. The SMD port is stable too
but those values I used is just for testing.
I have not used a true and calibrated SMD so I cannot properly
judge whether this is too much or too less sensitive.

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>