sprd_vsp: Mark the thread as real time; fix all the warnings
There is a piece of code there that elevates the priority of the current task
(the vsp kernel driver). It has been disabled, though.

Let's try enabling it but we will instead use the FIFO queue rather than the
Round Robin one. The reason is that we'd like it to have close to full
cpu power when it's turn to run.

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>