intelli_plug: Add back earlysuspend/lateresume callbacks
Yeah, yeah, earlysuspend has been depreciated and obsoletes but it's the
only thing that works well on the device, as of now.

You may disable the earlysuspend entirely; I have patches that allow that, but
the phone will not suspend.

And, for some reason, the fb notifier callback does not seem to be called
on screen off or on. So, add earlysuspend support back for now.