intelli_plug: Disallow jumps in nr_runs
One of the things I noticed in intelliplug v4.0(?) is that the hotplug
driver jumps from 4 to 1 to 4 to 1 and so on most of the time when the device is
idle. Though in v5.1 it does it less, it is still present.

I'm still researching why most of the "bursty" work happens during idle,
when there's one cpu online and it runs on the lowest frequency in the cpufreq table.
What I mean about "bursty" is that a high load is computed only for an iteration or
two.

On the governor side, interactive also exhibits this behavior, sometimes in sync with
intelliplug.
"From Idle, all cpus online while the governor ramps all cpu cores up to eleven"

I think what this need is some conservativeness or for it to have some time to distinguish
these bursts from normal high load.

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>