intelli_plug: Push current_profile selection out of the loop
The current_profile variable set during the computation of
the thread stats(? honestly I have no IDEA what the numbers mean...)

The variable is actually a function dependent of a user setting of how many
cores are allowed by the driver to be online simultaneously and also the
full nr profile setting when the previous aforementioned option is
set to the maximum value.
Even if it's not usually changed, there is a case where the user changed
either of the userland exposed settings while the loop is still running.
It can cause some illegal reads given that some profile/s are not
uniform in size.

Should we let it be or make the process one-time to further reduce the
likely of failure?

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>