Add ZEN_INTERACTIVE flag from Liquorix kernels
As the help says:
	Tunes the kernel simillar to Liquorix kernels.
	May increase both responsiveness and power usage, and decrease both latency and overall throughput.

This is the simillar config that is present on Liquorix kernel -- a custom kernel
which has a particular knack on performance. Though honestly the performance gained
is miniscule and most of changes reflect on the responsiveness of the machine.

The performance of the device is not that bad and
honestly that does not give any grounds for this patch to exist.
The only reason is for a record and makes is easier to revert this later on.

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>