mali-400: sc8830: Implement basic core scaling algorithm
The algorithm is intuitive enough that it scales up the number of enabled cores
in high loads then scales down as the load decreases.

But, the algorithm does this with a twist. Instead of iterating multiple of times
with the hope of it converging to the right scale, this attempts to derive
an absolute score, invariant of the frequency and the enabled cores.

This may work (actually it already does on my system) but it has a lot of cleaning to do.
This is also based on the arm_core_scaling.c of the default arm platform code.

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>