vmscan: Support multiple kswapd threads per node
TL;DR; Let's try out this patch in a blind test.

This is a patch written by Buddy Lumpkin <buddy.lumpkin@oracle.com>
and is marked as RFC. While being meant for much more recent kernel
version. This was pretty straightforward to backport to the 3.10
series.

You may read up the thread on some mailing list archive like:
https://lkml.org/lkml/2018/4/2/108
to learn more about the idea behind this patch about
increasing throughput.

Meanwhile the motive on our side is to leverage the flaw
in the hotplugging logic of sprdemand that it plugs in cpu cores
hoping that the processes may be migrated away to other
cores which, very likely, is not the case when kswapd runs
in either of the two reclaim modes.

Other reason is that it seems to allow, scanning the pages a bit
faster enabling the async/background kswapd to evict more pages
which in effect reduces the probability of doing a direct reclaim.

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>