mali-400: sc8830: Frequency awareness for the core scaling algorithm
The algorithm assumes that the frequency scales the capacity linearly
but in reality it really does not.

This patch extends the algorithm to scale the computed capacity given from the
performance differences between "adjacent" frequencies.

Fortunately since the mapping of the scale can be reversed, this patch took
the chance and extended the core scaling with frequency scaling.
The mali_core_freq_scale() now returns a suggested frequency.

By this patch, it should be also noticeable where this implementation fails.
The `utilization_gpu` value, whose this algorithm heavily relies on, is prone
to attenuation.

At maximum value of 256, we've basically lost all means of telling
how high the load is to the capacity; it could even mean that
the load is exactly and perfectly contained in the capacity
and raising the capacity will not do any performance increase.

On a side note, instead of the relative load score, the algorithm now tries to
predict the next `utilization_gpu` value and aims to maintain it in
a certain value. This works much better as a common base of comparisons given
that the relative load score is now relative to both the frequency
and the core count.

But because of the attenuation problem, the algorithm maintains it at around
205 and hope that it doesn't max out. This value can be tweaked by the
parameter `mali_core_tarutil`. The parameter accepts values from 0 to 256.
Lower values can help against utility spikes.

NOTE:
Also due to the attenuation problem, it is not possible to do some
large frequency or capacity jumps like the cpufreq ondemand governor does.
The best that the algorithm can do is ~20% capacity jump at a time
This is SLOW when compared how often the algorithm is executed.
This is made more noticeable when using the GPU from idle where
the capacities are at its lowest and the load would likely be
magnitudes higher than 20%, by comparison.

To alleviate this, `mali_core_minload` is a new parameter to set a minimum
capacity to scale down to. By default, the algorithm reserves 160
absolute load score. The parameter accepts values from 0 to 512.

This is still not enabled by default.

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>