mach-sc: Update the table of additional frequencies and its voltage
This is for the commit
"mach-sc: Allow more flexible changing of the frequency and voltage table for the processor"

After testing the frequencies extensively, there were still some lock ups due to the
voltage being set too low or the frequency being set to high. So I searched for the
highest and the lowest voltages to can be set and via some interpolation magic
I derived the values for the in between frequencies.

The highest possible frequency with a bit of a understandable stability issues
is at 1.45 GHz with the highest voltage at around 1.43V to 1.44V. Anything higher
is suicide.
I didn't find the lowest since the base load usually requires at least 200 MHz
to function. At around 100MHz or less, it is nearly impossible to start the whole
Android UI. I settled around 384MHz which is roughly in the middle of 200 and 400 MHz
with the lowest voltage sustainable voltage at 0.85V (actuall 0.83V).

This patch removes most close enough frequencies and tweaks the Voltages. In my tests,
this results on a much stable system, even using the UV entries with no crashes
even until now.

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>