sound, sprd: Fix #if directive checking for wrong CODEC sample rate macro
On the first condition CONFIG_SND_SOC_VBC_SRC_SAMPLE_RATE is being checked
but on the second condition CONFIG_SND_SOC_CODEC_SRC_SAMPLE_RATE is now
being checked.

Based on the code, one can deduce that the developer who wrote this
initially intended to have something to be checked whether they are
48000 or 32000 and have it fail when it's neither.

Fix the error when specifying 48000 as the value.

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>