android: lowmemorykiller: Cleanup Samsung enhancements
There's an additional code that slightly increases the min_score_adj
which tells the shrinker code which app should it reap first.

For some reason, this code block causes LMK to reap apps on the 2
(or 200 for newver versions) adj region
rather than reaping apps on the higher scores first.

This patch should expose a Kconfig option to prevent the code block
from being compiled to easily testing this issue.

But, I'm likely be removing it because of two reasons:
1. It violates the meaning of having an score_adj array
   ~ LMK's going to follow the swap usage anyways.
2. If ever we max out the swap after swapping out large amount of
   inactive pages, we also get large sum of free memory back in return.

   Now what would LMK do?
   With this code block active, min_adj_score would likely reach 0
   With the default behavior, min_adj_score will still depend on the free
   pages and pagecache pages.
3. I find this additional code is quite hideous, personally.

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>