mmc: sprd: Avoid calling might_sleep() when setting en/disabling the clock
We used the *prepare* versions to "prepare" a clock source, though the question
is that we don't have CONFIG_HAVE_CLK_PREPARE which mean we don't really have a
routine to prepare the clocks and the kernel code just uses might_sleep().

The Problem lies when all, if I counted everything correctly, calls to this
enable clock routine is in atomic context with either spinlock enabled or
irqs disabled. Then documentation specifically tells that
"Must not be called from within atomic context."
yet it is still called within it.

I think the clk_disable/enable functions should do the job, instead.

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>