input: sprd: headset: remove spinlocks in managing regulators
The spinlock makes the context atomic but the regulator API is actually
permitted to sleep so there's little meaning in maintaining a spinlock
when changing their states.

The regulators in questions are the main mic voltage,
its bias voltage (I think because the mic is a condenser mic)
and vbo(?)

Revert when something bad happens.

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>