Revert "arch, arm: sprd: Workaround compilation errors with BFS"
This reverts commit 2fd790bbed458a93010e036b16cf1d757557697c.

BFS wasn't even added on this branch, it causes instability enough
that wrecks the whole system as soon as it transitions to a deep sleep.

BFS pretty much prevents the system from waking from its deep sleep.

I made a change in commit 4c75eeacc737 to the mmc core specifically
to the part that handles the clocks but that was before testing BFS
so I don't know if that commit fixes the deep sleep issue or not.

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>