Revert "kanas: DONOTMERGE Disable most debug logs in the kernel build"
This reverts commit e8decd0792be732466f022dfee6dc17b4b495408.

I'm done checking the spinlock and mutexes uses across the device.
I don't know, honestly, what have changed since the kernel still
works functionally and behaviorally the same as before...

The changes I did are precautionary measures that the kernel devs
said to avoid complications with preemption.

Still, linux-rt patchset will not work with these changes as it fails
after importing the rt-mutexes part which should be around the middle
of the 400+ patches. darn.

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>