cpufreq: sprdemand: use raw_smp_processor_id instead of smp_processor_id
One of the changes in sprdemand is to only allow CPU0 to change the frequency
and plug in/out cores, I think.

Then, it uses smp_processor_id() which disables preemption as stated in the
kernel documentation "Unreliable Guide To Hacking The Linux Kernel" chapter 6.

Though it's not clear why, but the timer context is, AFAIK, not an
atomic context but, in other words, is a preemptible region. After checking the
cpu id or number, there no guarantee that the id we get is still the same
id of the cpu executing the rest of the code. It may be unlikely but it is
possible.

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>