mmc: bus: Resume bus before shutting it down
Due to MMC_BLOCK_DEFERRED_RESUME feature to defer the resume of the mmc bus
(which could easily take hundreds of milliseconds). This would probably cause
the suspended bus to ignore commands causing a timeout on shutting it down
and when done on poweroff will cause a kernel BUG and a kernel panic.

This can easily be solved by resuming the bus before shutting it down or by
not using this feature at all.

Issue and test: On the LPM mode (something like powered-off charging) of SM-G355H,
when the phone is disconnected from the charger, the LCD will power on for
a brief moment then fades to white.
The phone will be unresponsive and will also start to heat up
until it's forcibly rebooted.

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>