mmc: block: sprd: Add a shutdown callback to finish all reqs in the queue
This slightly modifies the mmc queue suspend routine where the queue
is emptied (or executed) before suspending as inidicated by an extra parameter
wait.

This doesn't really have any functional improvement and the
file integrity improvement aspect is yet to be tested since this forces
all queued up IO to be done when shutting down the mmc device.

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>