mmc: core: Only allow sd cards to be deferred resumed; copy paste mmc_resume_host() in mmc_resume_bus()
The real issue is that on SD cards the struct bio info for all partitions of the
SD card becomes incorrect causing every sub/systems depending on it to fail
(dm,dm-crypt,f2fs,exfat,sd card apps)

Reading unto it causes numerous EOD and every read tricks the kernel into
believing that we have a bad sector somewhere. Though on the bright side,
all writes are blocked meaning we don't really lose any saved data.
Can't say for sure about unsaved data.

For now, I'm trying to trace back the resume path and it seems that the
detect function fires too early before the hardware in question truly powers up.
That's the hypothesis since that the detect() is removed by this commit.
Or, it could also be because the emmc cards are not deferred resumed.

I may jinx the situation if I report how it went after 3 hours of testing.

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>