mmc: core: also defer detection on defered resume.
The original patchset calls the detect() callback with 0 jiffy of delay.
But, it seems that the SD card host needs a bit more time to "wake up"
despite being fully powered.

This is the fix to the hypothesis that detect() can misdetect
that the card has been removed or generally cause wrong updates.

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>