mmc: block: sprd: Ignore reset_done on resetting on certain occasions
In some cases, a single reset doesn't work for some weird reason.
One case of this is when the eMMC gets CRC errors when reading a block.
The driver then attempts to reset the block device (which succeeds)
before retrying the command only to fail again.

The driver will not reset it again on this case which will render the
device nearly inoperable. If we did try to reset it second
time then retry, it should miraculously continue on as if nothing happened.

Obviously this is not a solution but a work around because if CRC errors
is the cause of it triggering the read error, the underlying data
might have been corrupted. This WA should at least let a dev
check what's going on.

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>