mali-400: sc8830: Remove undefined scaling scenario; Fix DFS related deadlock
In one of the "tricks" in the core scaler, under the "old_freq == max_freq" condition
the statement is commented out. This is the case where the load is above 230, the algorithm
rounds it up to the full capacity of the current core count. The problem with this is that it
can help cause some oscilation between core counts which is somewhat undesirable. Having the
frequency oscillate with the same core count is better than the former which is why it is
removed.

On commit 6e032df18a72b592abfb2f6d267798fd7d4a7e74(Revert "genirq: Set irq thread to RT priority on creation")
The hangs observed there seems to be a deadlock within the mali platform code and the experimental
core scheduler. With the patch "genirq: Set irq thread to RT priority on creation" applied,
it makes it more likely to occur. The hang will put mainly graphics using apps into an uninterruptible
sleep (via an IOCTL, I suspect) and thus cannot be killed, even reboot causes a timeout in the mali driver.

Using sysrq I got these suspicously looking logs:
6,57073,10110875046,-;[c0] SysRq : Show Blocked State
6,57074,10110875106,-;[c0]   task                PC stack   pid father
6,57075,10110875177,-;[c0] kworker/1:1     D c0747d2c     0    30      2 0x00000000
6,57076,10110875246,-;[c0] Workqueue: pm pm_runtime_work
6,57077,10110875319,-;[c0] [<c0747d2c>] (schedule+0xc70/0xd90) from [<c0745614>] (schedule_timeout+0x2c/0x298)
6,57078,10110875369,-;[c0] [<c0745614>] (schedule_timeout+0x2c/0x298) from [<c07469e8>] (__down+0x70/0xa8)
6,57079,10110875423,-;[c0] [<c07469e8>] (__down+0x70/0xa8) from [<c006052c>] (down+0x3c/0x4c)
6,57080,10110875480,-;[c0] [<c006052c>] (down+0x3c/0x4c) from [<c0383070>] (mali_platform_power_mode_change+0x24/0x214)
6,57081,10110875537,-;[c0] [<c0383070>] (mali_platform_power_mode_change+0x24/0x214) from [<c0378f68>] (mali_pm_runtime_suspend+0x50/0x64)
6,57082,10110875592,-;[c0] [<c0378f68>] (mali_pm_runtime_suspend+0x50/0x64) from [<c0361a2c>] (mali_driver_runtime_suspend+0x18/0x20)
6,57083,10110875651,-;[c0] [<c0361a2c>] (mali_driver_runtime_suspend+0x18/0x20) from [<c038e65c>] (pm_generic_runtime_suspend+0x3c/0x48)
6,57084,10110875705,-;[c0] [<c038e65c>] (pm_generic_runtime_suspend+0x3c/0x48) from [<c0392e7c>] (__rpm_callback+0x3c/0x6c)
6,57085,10110875755,-;[c0] [<c0392e7c>] (__rpm_callback+0x3c/0x6c) from [<c0392f2c>] (rpm_callback+0x80/0x90)
6,57086,10110875803,-;[c0] [<c0392f2c>] (rpm_callback+0x80/0x90) from [<c0393c68>] (rpm_suspend+0x318/0x570)
6,57087,10110875852,-;[c0] [<c0393c68>] (rpm_suspend+0x318/0x570) from [<c0394938>] (pm_runtime_work+0x84/0x9c)
6,57088,10110875907,-;[c0] [<c0394938>] (pm_runtime_work+0x84/0x9c) from [<c0053c00>] (process_one_work+0x280/0x504)
6,57089,10110875960,-;[c0] [<c0053c00>] (process_one_work+0x280/0x504) from [<c0054e30>] (worker_thread+0x1c4/0x354)
6,57090,10110876013,-;[c0] [<c0054e30>] (worker_thread+0x1c4/0x354) from [<c005a96c>] (kthread+0xb0/0xbc)
6,57091,10110876065,-;[c0] [<c005a96c>] (kthread+0xb0/0xbc) from [<c000f890>] (ret_from_fork+0x14/0x20)
...
6,57144,10110879739,-;[c0] kworker/u8:6    D c0747d2c     0 13900      2 0x00000000
6,57145,10110879789,-;[c0] Workqueue: gpu_dfs gpu_dfs_func
6,57146,10110879844,-;[c0] [<c0747d2c>] (schedule+0xc70/0xd90) from [<c03931c4>] (__pm_runtime_barrier+0xf0/0x114)
6,57147,10110879895,-;[c0] [<c03931c4>] (__pm_runtime_barrier+0xf0/0x114) from [<c03945d0>] (pm_runtime_barrier+0x70/0xc0)
6,57148,10110879947,-;[c0] [<c03945d0>] (pm_runtime_barrier+0x70/0xc0) from [<c03820e4>] (_mali_osk_pm_dev_barrier+0x24/0x2c)
6,57149,10110880002,-;[c0] [<c03820e4>] (_mali_osk_pm_dev_barrier+0x24/0x2c) from [<c0373a60>] (mali_pp_scheduler_set_perf_level+0x9c/0x76c)
6,57150,10110880056,-;[c0] [<c0373a60>] (mali_pp_scheduler_set_perf_level+0x9c/0x76c) from [<c0382278>] (mali_perf_set_num_pp_cores+0x1c/0x20)
6,57151,10110880109,-;[c0] [<c0382278>] (mali_perf_set_num_pp_cores+0x1c/0x20) from [<c03834c4>] (mali_core_freq_quick_set_saved+0x30/0x38)
6,57152,10110880160,-;[c0] [<c03834c4>] (mali_core_freq_quick_set_saved+0x30/0x38) from [<c03822dc>] (gpu_dfs_func+0x40/0x164)
6,57153,10110880212,-;[c0] [<c03822dc>] (gpu_dfs_func+0x40/0x164) from [<c0053c00>] (process_one_work+0x280/0x504)
6,57154,10110880266,-;[c0] [<c0053c00>] (process_one_work+0x280/0x504) from [<c0054e30>] (worker_thread+0x1c4/0x354)
6,57155,10110880317,-;[c0] [<c0054e30>] (worker_thread+0x1c4/0x354) from [<c005a96c>] (kthread+0xb0/0xbc)
6,57156,10110880368,-;[c0] [<c005a96c>] (kthread+0xb0/0xbc) from [<c000f890>] (ret_from_fork+0x14/0x20)

It turns out that the platform code is running the DFS in a separate thread, a semaphore lock is taken,
and calls to Mali to enable/disable cores. The internal code in the mali driver ends up calling the
mali_platform_power_mode_change() function which is also in the platform code and incidentally also
takes the same semaphore lock. These whole process is reentrant which means it's a deadlock.

There are two implemented ways to enable/disable cores in the experimental core scaler:
a workqueue and a direct call method. Obviously, that deadlock occurs with the direct call method.

The solution? Just use the workqueue method, remove the direct call method.

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>