sprd: vibrator: Fix locked vibration and wrong output time
This is a partial rewrite based on the newer vibrator.c source
(which resides in the drivers/platform/sprd location). Again, I said "based".

Vibration based on the power regulators, usually the fixed-pulse vibration
thingy, can suffer the feature of the regulator_enable function.
This function can be called multiple times in a row but IN A CONDITION
that regulator_disable must be called by the same number of times.
Fortunately, you can call regulator_disable multiple times and still
need a single call to regulator_enable to enable the regulator.

The problem is that when the 'enable' file is written in a quick succession
or at least written before the timer runs out. Chances are that the
regulator_enable() will be called twice yet the regulator_disable()
will only called once. This is due to the fact that only the timer callback
can call it and it won't do it twice in a single timeout event.

This also had a problem of outputing nanoseconds when reading the
'enable' file which doesn't make sense. You write milliseconds in
and get nanoseconds out. Why is that?

This patch will rewrite the whole source for readability purposes and to
avoid using the global workqueue. The newer source still has the workqueue
and the schedule_work(). I have removed it and just directly call the function
when needed. schedule_work() has the tendency to block quick successive writes
(something like a debouncer) but we don't actually need it.

Not using the workqueue should speed things up (in the nanosecond scale).

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>