cpufreq: sprdemand: Fix a certain crash when the timer function executes while deinitializing the governor
In an unlikely event where sd_dbs_timer() will runs a bit after freeing the private
governor data for sprdemand, we can get a kernel panic due to a NULL pointer
dereference.

Reproducing this could be quite a challenge given people don't change governors
in a drop of a hat. But, it can still happen while switching out governors.

Signed-off-by: Mark Enriquez <enriquezmark36@gmail.com>